+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.1.7] - 2025-07-10
+
+- Renamed dynamic-link certificate choice to device-link certificate choice.
+- Updated certificate choice endpoint to use /device-link/ paths.
+- Added `initialCallbackURL` field with regex validation.
+- Added `deviceLinkBase` to session response.
+
 ## [3.1.6] - 2025-07-08
 
 ### Added